requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
openpyxl
statsmodels
prophet
scipy
joblib
requests  # Added for API calls
plotly>=5.18.0
nbformat>=5.9.2
beautifulsoup4>=4.9.3
python-dateutil>=2.8.0  # For date handling in hexagonal architecture